requirements.txt
google-adk==1.14.0
pandas==2.2.2
python-dotenv==1.0.1
